fix: rename impl files to {module}_impl.h/.cpp (kv_module_impl, not kv_impl)
